+++ apps/api/src/events/events.service.ts
@@ -10,13 +10,12 @@ import { PrismaService } from '../prisma/prisma.service';
 import { StripeService } from '../stripe/stripe.service';
 import { S3Service } from '../cloud/cloud.service';
 import { CreateEventDto } from './dto/create-events.dto';
-import { UpdateEventDto } from './dto/updtate-event.dto';
-import { FormPayLoad } from './dto/create-form.dto';
+import { UpdateEventDto } from './dto/update-event.dto';
 import { GetEventByOrgDto, GetEventByOrgIdDto } from './dto/get-events.dto';
 import excludeKey from '../utils/exclude';
-import {Events, User} from '@prisma/client';
-import {EventEmitter2, OnEvent} from "@nestjs/event-emitter";
-import {AiService} from "../ai/ai.service";
+import { Events, User } from '@prisma/client';
+import { EventEmitter2, OnEvent } from '@nestjs/event-emitter';
+import { AiService } from '../ai/ai.service';
 
 @Injectable()
 export class EventsService {
@@ -130,7 +129,7 @@ export class EventsService {
             });
 
             if (!event) {
-                throw new NotFoundException("Event doesn't exist")
+                throw new NotFoundException("Event doesn't exist");
             }
 
             const data = await this.prismaService.events.update({
@@ -175,12 +174,12 @@ export class EventsService {
                     message: e.message,
                 });
             } else if (e instanceof UnprocessableEntityException) {
-                throw  new UnprocessableEntityException({
+                throw new UnprocessableEntityException({
                     ok: false,
                     message: e.message,
                 });
             } else {
-                throw  new InternalServerErrorException({
+                throw new InternalServerErrorException({
                     error: e,
                 });
             }
@@ -273,7 +272,7 @@ export class EventsService {
         } catch (e) {
             // Use exception filters to handle exceptions and return appropriate responses
             if (e instanceof NotFoundException) {
-                throw  new NotFoundException({
+                throw new NotFoundException({
                     ok: false,
                     message: e.message,
                 });
@@ -371,12 +370,12 @@ export class EventsService {
                 });
 
                 if (!data) {
-                    throw  new NotFoundException();
+                    throw new NotFoundException();
                 }
 
                 return {
                     ok: true,
-                    message: 'Event registerd successfully',
+                    message: 'Event register successfully',
                 };
             }
         } catch (e) {
@@ -385,7 +384,7 @@ export class EventsService {
             } else if (e instanceof ServiceUnavailableException) {
                 throw new ServiceUnavailableException();
             } else if (e instanceof ConflictException) {
-                throw new ConflictException('user already registerd');
+                throw new ConflictException('user already registered');
             } else {
                 throw e;
             }
@@ -467,48 +466,6 @@ export class EventsService {
         }
     }
 
-    async createForm(data: FormPayLoad) {
-        try {
-            const event = await this.prismaService.events.findUnique({
-                where: {
-                    organizationId: data.organizationId,
-                    id: data.eventId,
-                },
-            });
-
-            if (!event) {
-                throw new NotFoundException();
-            }
-
-            const formSchema = await this.prismaService.field.create({
-                data: {
-                    label: data.data.label,
-                    placeholder: data.data.placeholder,
-                    required: data.data.required,
-                    type: data.data.type,
-                    options: data.data.options,
-                    Events: {
-                        connect: {
-                            id: data.eventId,
-                        },
-                    },
-                },
-            });
-
-            return {
-                ok: true,
-                message: 'schema updated successfully',
-                data: formSchema,
-            };
-        } catch (e) {
-            if (e instanceof NotFoundException) {
-                throw new NotFoundException();
-            } else {
-                throw e;
-            }
-        }
-    }
-
     async uploadEventCover(file: Express.Multer.File, eventSlug: string) {
         try {
             const publicUrl = await this.cloudService.uploadFile(file);
@@ -536,35 +493,6 @@ export class EventsService {
         } catch (e) {
             if (e instanceof InternalServerErrorException) {
                 throw new InternalServerErrorException();
-            }
-        }
-    }
-
-    async getEventFormScheme(eventSlug: string) {
-        try {
-            const eventSchema = await this.prismaService.events.findUnique({
-                where: {
-                    slug: eventSlug,
-                },
-                select: {
-                    form: true,
-                },
-            });
-
-            if (!eventSchema) {
-                throw new NotFoundException();
-            }
-
-            return {
-                ok: true,
-                message: 'Event schema found',
-                data: eventSchema.form,
-            };
-        } catch (e) {
-            if (e instanceof NotFoundException) {
-                throw new NotFoundException();
-            } else {
-                throw e;
             }
         }
     }
@@ -602,124 +530,6 @@ export class EventsService {
         }
     }
 
-    async toggleFormPublishStatus(eventSlug: string, shouldPublish: boolean) {
-        try {
-            const event = await this.prismaService.events.findUnique({
-                where: {
-                    slug: eventSlug,
-                },
-                select: {
-                    form: true,
-                },
-            });
-            if (!event) {
-                throw new NotFoundException();
-            }
-
-            if (!event.form.length) {
-                throw new InternalServerErrorException();
-            }
-
-            const newEventStatus = await this.prismaService.events.update({
-                where: {
-                    slug: eventSlug,
-                },
-                data: {
-                    isFormPublished: shouldPublish,
-                },
-            });
-
-            return {
-                ok: 'true',
-                message: 'form status was changed',
-                data: newEventStatus,
-            };
-        } catch (e) {
-            if (e instanceof InternalServerErrorException) {
-                throw new InternalServerErrorException({
-                    message: 'Please create a form schema to publish the form',
-                });
-            }
-            if (e instanceof NotFoundException) {
-                throw new NotFoundException();
-            }
-
-            return e;
-        }
-    }
-
-    public async addUserFormSubmission(
-        info: Record<string, string | Array<string>>,
-        eventId: string,
-        userId: string,
-    ) {
-        const event = await this.getEventById(eventId);
-        if (!event) return new NotFoundException();
-
-        await this.prismaService.response.create({
-            data: {
-                data: info,
-                user: {
-                    connect: {
-                        uid: userId,
-                    },
-                },
-                Events: {
-                    connect: {
-                        id: event.id,
-                    },
-                },
-            },
-        });
-    }
-
-    async getregisterParticipantsFormSubmissions(id: string, userId: string) {
-        try {
-            const schema = await this.prismaService.events.findUnique({
-                where: {
-                    slug: id,
-                },
-                select: {
-                    form: true,
-                    id: true,
-                },
-            });
-
-            const label = {};
-            schema.form.forEach((el) => {
-                label[el.id] = el.label;
-            });
-
-            const data = await this.prismaService.response.findMany({
-                where: {
-                    userUid: userId,
-                    eventsId: schema.id,
-                },
-            });
-            if (!data) return new NotFoundException();
-
-            const result = {};
-
-            schema.form.forEach((el) => {
-                // @ts-ignore
-                if (el.id in data[0].data) {
-                    result[el.label] = data[0].data[el.id];
-                }
-            });
-
-            return {
-                ok: true,
-                data: result,
-                message: 'recieved form successfully',
-            };
-        } catch (e) {
-            if (e instanceof NotFoundException) {
-                throw new NotFoundException();
-            } else {
-                throw e;
-            }
-        }
-    }
     async deleteEvent(eventId: string) {
         try {
             const event = await this.getEventById(eventId);
@@ -854,10 +664,14 @@ export class EventsService {
 
     @OnEvent('event.updated')
     async handleEventUpdated(event: Events) {
-        const embedDescription = await this.AiService.generateEmbedding(JSON.stringify(event.description));
+        const embedDescription = await this.AiService.generateEmbedding(
+            JSON.stringify(event.description),
+        );
         const embedName = await this.AiService.generateEmbedding(JSON.stringify(event.name));
-        await this.prismaService.$queryRaw`UPDATE public."Events" SET embedding_description = ${embedDescription} WHERE id = ${event.id}`;
-        await this.prismaService.$queryRaw`UPDATE public."Events" SET embedding_title = ${embedName} WHERE id = ${event.id}`;
+        await this.prismaService
+            .$queryRaw`UPDATE public."Events" SET embedding_description = ${embedDescription} WHERE id = ${event.id}`;
+        await this.prismaService
+            .$queryRaw`UPDATE public."Events" SET embedding_title = ${embedName} WHERE id = ${event.id}`;
     }
 }
